refactor(cli): adopt resource-based command structure
