+++ README.md
@@ -152,13 +152,12 @@ VERSION:
    v0.1.0
 
 COMMANDS:
-   init, i            Initialize a Tempo project
-   define, d          Define templates and actions for component or variant
-   new, n             Generate a component or variant based on defined templates
-   sync, s            Process & sync asset files into component templates
-   register, r        Register is used to extend tempo
-   clean              Remove temporary metadata generated by tempo run
-   help, h            Shows a list of commands or help for one command
+   init            Initialize a Tempo project
+   component       Define templates and actions for component or variant
+   variant         Generate a component or variant based on defined templates
+   register        Register is used to extend tempo
+   sync            Process & sync asset files into component templates
+   help            Shows a list of commands or help for one command
 
 GLOBAL OPTIONS:
    --help, -h     show help
@@ -250,6 +249,35 @@ This command shares all the flags from `new component`, plus:
 </dl>
 </details>
 
+### register
+
+Register is used to extend tempo.
+
+```bash
+tempo register -h
+tempo register -h functions
+```
+
+- **functions** – Register a function provider from a local go module path or a remote repository
+
+```bash
+tempo register functions --name sprig --url https://github.com/indaco/tempo-provider-sprig.git
+```
+
+<details>
+<summary><strong>Functions Flags</strong> (<code>tempo register functions</code>)</summary>
+<dl>
+  <dt><code>--name</code> (<code>-n</code>) <em>value</em></dt>
+  <dd>Name for the function provider</dd>
+
+  <dt><code>--url</code> (<code>-u</code>) <em>value</em></dt>
+  <dd>Repository URL</dd>
+
+  <dt><code>--path</code> (<code>-p</code>) <em>value</em></dt>
+  <dd>Path to a local go module provider</dd>
+</dl>
+</details>
+
 ### sync
 
 Process & sync asset files into templ component files
@@ -266,11 +294,9 @@ This command scans the `input` folder for CSS and JS files and injects their con
 
 Whenever you update your CSS or JS files, simply run `tempo sync` to propagate the changes.
 
-> [!TIP]
-> **Live Reload:** If you're using a live reload tool like [air](https://github.com/air-verse/air), [templier](https://github.com/romshark/templier), or [watchexec](https://github.com/watchexec/watchexec), pass `--summary none` to reduce unnecessary output.
+> [!TIP] > **Live Reload:** If you're using a live reload tool like [air](https://github.com/air-verse/air), [templier](https://github.com/romshark/templier), or [watchexec](https://github.com/watchexec/watchexec), pass `--summary none` to reduce unnecessary output.
 
-> [!NOTE]
-> **Want to use tempo without scaffolding?**
+> [!NOTE] > **Want to use tempo without scaffolding?**
 > Check out [Using tempo sync as a Standalone Command](#-using-tempo-sync-as-a-standalone-command).
 
 <details>
@@ -304,46 +330,6 @@ Whenever you update your CSS or JS files, simply run `tempo sync` to propagate t
   <dd>Export the summary to a JSON file</dd>
 </dl>
 </details>
-
-### register
-
-Register is used to extend tempo.
-
-```bash
-tempo register -h
-tempo register -h functions
-```
-
-- **functions** – Register a function provider from a local go module path or a remote repository
-
-```bash
-tempo register functions --name sprig --url https://github.com/indaco/tempo-provider-sprig.git
-```
-
-<details>
-<summary><strong>Functions Flags</strong> (<code>tempo register functions</code>)</summary>
-<dl>
-  <dt><code>--name</code> (<code>-n</code>) <em>value</em></dt>
-  <dd>Name for the function provider</dd>
-
-  <dt><code>--url</code> (<code>-u</code>) <em>value</em></dt>
-  <dd>Repository URL</dd>
-
-  <dt><code>--path</code> (<code>-p</code>) <em>value</em></dt>
-  <dd>Path to a local go module provider</dd>
-</dl>
-</details>
-
-### clean
-
-Remove temporary metadata generated by `tempo sync`
-
-```bash
-tempo clean
-```
-
-- Deletes `.tempo-lastrun`, a file used to track asset processing timestamps.
-- This command is useful if you need to reset the sync state before running `tempo sync` again.
 
 ## ⚡ Using tempo sync as a Standalone Command
